requirements.txt
python-dotenv~=1.0.0
loguru==0.7.2
discord.py==2.3.2
pydantic==2.5.3
asyncpg==0.29.0
SQLAlchemy==2.0.25
sqlalchemy[asyncio]
alembic==1.13.1
pynacl==1.5.0
aiohttp~=3.9.1
redis==5.0.1
orjson==3.9.12
apscheduler==3.10.4